Refactor FocusLauncherForm to use gap prop for improved layout and readability

+++ increaser/api/projects/resolvers/deleteProject.ts
@@ -1,12 +1,17 @@
 import { assertUserId } from '../../auth/assertUserId'
 import * as projectsDb from '@increaser/db/project'
 import { ApiResolver } from '../../resolvers/ApiResolver'
+import { couldProjectBeDeleted } from '@increaser/entities-utils/project/couldProjectBeDeleted'
 
 export const deleteProject: ApiResolver<'deleteProject'> = async ({
   input: { id },
   context,
 }) => {
   const userId = assertUserId(context)
+
+  if (!couldProjectBeDeleted(id)) {
+    throw new Error(`A project with id ${id} could not be deleted.`)
+  }
 
   await projectsDb.deleteProject(userId, id)
 }
